Defer ensuring visibility on enter/exit split-screen
Windowing mode changes also ensure activity visibility and config
after the change, so recursive calls may do this too early and
cause redundant configuration change/relaunch messages sent to
client.
For example, when a fullscreen non-resizeable activity is launched
when split-screen stack is active, then split-screen stacks are
also becoming fullscreen and get a new configuration. But this
shouldn't happen because they are about to be occluded by the newly
launched activity.

Bug: 74375006
Test: ActivityLifecycleTests
Change-Id: I6b701fba40f9b66873d0ec12201b4e8c5fd3ffa6

Former-commit-id: 9da138a349858bac53a25a1a09f137c2e8d6932a