Keyguard: fix possible divergent state
Fixes an issue where the state in SystemUI and AM/WM could
get out of sync: If the keyguardGoingAway transition is
started and then later canceled by SystemUI, it never
tells AM that it canceled it.
This causes SystemUI to show itself in keyguard mode,
meanwhile AM/WM do not hide the underlying app because
the status bar is in the going away transition as far
as it is concerned.

Change-Id: I5966f0655fd694ae599a83dda7a9f6463088632b
Fixes: 76367931
Test: delay keyguard going away callback, unlock phone, during the delay, remove SIM; verify it does not end up in bad state

Former-commit-id: 6ec76b728871407cf19fac4bd802e1b0eb3a2a30