Use PrinterHashMap in MDNSFilteredDiscovery
Before MDNSFilteredDiscovery used the ip-addresses to remove printers,
but this did not work as once the nsd service is lost, you cannot resolve
it anymore. Hence do the same thing as the other plugins, and remove a
lost service if the name matches.

Also do not try back when the callbacks are already gone.

Fixes: 70622095
Test: Removed Wifi when print service recommendations were shown and saw
      all recommendations (including samsung) to go away. Without this
      change the samsung service never realized when printers went away.

Change-Id: I7371ef00d626738741d39a2f3669497bfc9fd080

Former-commit-id: 8f65cc365f0010473e3273d7906c7e1738c448ab