Use Empty List for Default PhysicalChannelConfig
If a caller registers to listen() for a physical
channel config message, the beavior is to immediately
trigger an unsol. In the case that a list hasn't been
sent from the modem, we should send a default empty
list rather than a null. This is the preferred
approach for Android APIs because it reduces the
reliance on null checks by callers.

Bug: 78536847
Test: ¯\_(ツ)_/¯
Change-Id: I9e59195f5d64264ab473b2fd7e4f7eb6446c4829

Former-commit-id: c9bad6ef7aa2448c58d94791032c0b66fdbd9120