Fixes non-matching instant app behavior
This change ensures that a correctly formatted intent is sent to the
instant app installer when no match can be found. Prior to this change,
the original intent with the installer set as the component would be
sent. After, we send an install intent with a sanitized intent set as an
extra, as expected.

Bug: 63117034
Test: manual - launch non-matching intents with flag set
Change-Id: I70a3e2fab55e8f0524bbc99e88fba98dc2e28aaa

Former-commit-id: 3e0b8ee39ab3435abd88c03069272da6e605b6a0