Don't allow non-showOnLock activity to be next on finish.
Previously, we required the next activity (determined by top running
activity) to be showOnLock in the case the keyguard was locked and
the current activity is finished. However, this was bypassed if the
focused stack had a top running activity.

This changelist applies the same restrictions to the aforementioned
activity.

Test: atest ActivityStackSupervisorTests#testTopRunningActivity
Change-Id: I70ae64e0b0f4765b383bce91017bb675d0376d31
Fixes: 76424176

Former-commit-id: 86918964d693e3843858ebd2e3c08df5c1b2adc6