Use dedicated thread for sync manager
We get WTF indicating that a message posted onto the BG handler thread wasn't
processed in 10 minutes...

It's either there's some race in SyncJobService, or sometimes the BG thread
is actually very crowded and a message won't be delivered in 10 minutes.

Both are hard to think but let's assign a different thread to exclude the
later hypothesis.

Bug: 78590632
Test: boot and make sure various syncs run.
Change-Id: If63d03885ba0d126a5aac196d3b8a2bc7bd2525d

Former-commit-id: e08a5c21688416cb07ae1dfda8009dd792d809ea