Make isValidSystemNonAuxAsciiCapableIme stateless
This CL simplifies SettingsLib by making
isValidSystemNonAuxAsciiCapableIme() method stateless.

Although this can be a minor user-visible behavior change, we believe
it is acceptable because the behavior we are going to drop is a
fallback rule in case there is no system IME that could be considered
ASCII capable without actual AsciiCapable attribute..  Even without
this fallback rule the System Settings does not fall into a bad state.
It only affects the sort order of the IME enabler page and minor
optional cosmetic behavior.

The bottom line is that if a pre-installed IME is ASCII capable, it
really should annotate itself to be so, rather than relying on some
heuristic logic that we have had.

With this CL, SettingsLib no longer depends on InputMethodUtils.

Fix: 77730201
Test: atest InputMethodPreferenceTest
Test: make RunSettingsLibRoboTests -j
Change-Id: I1352bf42dcb9219a370a16a381ea39c43c57637a

Former-commit-id: 4faac7d9ef6e61ea32e3fa55dd18c35bcc525ac6