Reset auto-brightness on new BrightnessConfiguration
In theory, the newly pushed configuration should take into account all
of the brightness adjustments that the user has made throughout the day,
so applying any part of the short term model on top of that would be
effectively adjusting for them twice over. To avoid getting into this
state, we reset the short-term model.

Bug: 78910182
Test: Manual
Change-Id: I52121eac84eb8ceb914a94a95af2f174ab2f5b73

Former-commit-id: b8f4f77d9e12e125408b058bcbd505f4a0bdd50e