Fix issue #77931346: The notification that should not be named appeared
Turn off the notification.  This just adds a boolean to
conditionally turn off the generation of the notification...
all of the logic for what is in it and when it is shown is
still there for now, to reduce the amount of change at this
point, continue to have it in bug reports, and I am not sure
if we might want to do some other visualization of this in
the future.

Bug: 77931346
Test: manual
Change-Id: I38663b6defb15ae925c2e6a4df9e2f0a9a023050

Former-commit-id: 070d142993403cc2c42eca808ff3fafcee220ac4