Allocate new native handle to store sidebandStream
When we get sidebandStream from TvInput, we only store the address of
native handle, the address point to native_handle_t in _hidl_replay.
After openStream is finished, the _hidl_replay will be freed
automatically, therefore we will use the freed memory in
setSidebandStrem, It will cause the writeNativeHandle fail in
setSidebandStrem.

Change-Id: Idd2e8343670ef5a3048836f1db978cd7c04378b7

Former-commit-id: 4237ed70b2a41b585f79f2b5d12e96193743342d