Support generic filters when EXTERNAL flag is set
This change opts instant app resolution out of hash prefix matching if
the intent filter being tested does not define an authority only when
the MATCH_EXTERNAL flag is set. This allows for a class of instant apps
that generically handle certain mime types or actions, but aren't tied
to a particular authority / domain.

This also moves the null check for resolution to a higher level,
accounting for connection-related failures to ensure that the installer
always gets a properly formatted intent.

Bug: 63117034
Test: manual - force connection timeouts and observe correct install intent
Test: manual - launch a generic type handler for a made up mime type;
               observe correct resolution
Test: atest EphemeralTest
Change-Id: I0101e3219f543897a488ddd3d7c4c884f7d6cac3

Former-commit-id: a6862c238039cba1b2f6175beca58f8d76e1bef9