[Android WebView] print minimum sdk version required for WebView APKs
'adb shell dumpsys webviewupdate' currently prints whether each WebView
APK on the device is a valid WebView implementation. if the APK isn't
valid, that command also prints the reason why it isn't.
A WebView APK might have too low of a targetSdkVersion to be a WebView
implementation (for the current device), in such cases it would be nice
(for debugging reasons) to know what the required sdk version is.

This CL makes the command 'adb shell dumpsys webviewupdate' print the
targetSdkVersion required for an APK to be a WebView implementation.

Bug: 79236693
Test: run 'adb shell dumpsys webviewupdate' to ensure the current
minimum target-sdk version is printed.

Change-Id: I55e13611c6685c3185addd4069704e0ff8b514f6

Former-commit-id: 6823c48b8b85e2b198f54670ad491b647d2ad330