Revert Nav bar contrast under magnification
I don't have an answer for b/74464336, which is caused by a
race between the nav bar becoming opaque and the content
behind the nav bar zooming past it. The ideal solution -
to exclude the nav bar's background from magnification -
is not feasible right now because magnification is
performed on complete windows.

I think rolling back to O is the least bad thing here.

Bug: 74464336
Test: Verify that magnification works the way it did in O.
Change-Id: I3b4028ddba6a3d8541948dd89813728c347ad30f

Former-commit-id: b7f2c9cf291b1807a296f36e751d56c8c89df772