Allow ImageDecoder to read SkFILEStream directly
Bug: 78866720
Test: Manual

If the input stream is seekable, SkGifCodec will read from it directly.
If not, it has to copy (parts of) the input in order re-decode when it
loops. Directly use the seekable SkFILEStream so that SkGifCodec skips
the copy, saving memory proportional to the size of the file.

Depends on a change in upstream Skia which allows SkFILEStream to treat
the initial offset as the beginning of the file:
https://skia-review.googlesource.com/c/skia/+/126511

Change-Id: Iefb58785157ba684ad3603778175b3dba97567b2

Former-commit-id: 0c699add019cde2dc32a5f090b20857fc1c753f2