Fixes docked stack from being invisible from assistant stack over
Set the special case for fullscreen assistant over docked stack in
minimized mode for the docked app to be visible. This would also affect
the dividerview to animate minimized when assistant is over. If the
divider view animation had not finished while swipe up from fullscreen
assistant, the divider not be minimized because the animation finishes
and was not stopped and set back to minimized, therefore cancel the
animation when divider controller tells divider view to minimize.

Test: docked something, minimized, invoke assistant, tap to make it
fullscreen, then swipe up
Fixes: 74551860
Change-Id: If7081fd642db8f15291a19573fd449fde5440893

Former-commit-id: a51dcaae7a0510dc8db84b93673227c21668fd34