Greylist cloned SDK fields
Doclava will expose a field from a hidden interface if it is inherited
by an SDK class by temporarily creating a new FieldInfo object which
pretends the field is in fact declared in the SDK class.

This mechanism was missed by the routine which prints non-SDK APIs
and the fields are considered non-SDK. Until this is fixed, put the
affected fields on the light greylist to unblock CTS tests.

Bug: 64382372
Bug: 73082683
Test: atest CtsCurrentApiSignatureTestCases
Change-Id: Ic10d26b82ad8f324280e0198f1d3e8b0ab824d35

Former-commit-id: dffafa5b66939e12c2eb4e8b5867b4bd82e2400c