[Notif] Fix a11y when guts is open
The content in the back is showing while guts is up since we never make
the content invisible. To avoid making a semi-risky change before the
GAR cutoff, we're listening to guts open/close and toggling the content
views' important for accessibility values.

After guts is closed, this will also re-focus the notification that the
user had long-pressed instead of leaving them without a target/making
them lose track.

Fixes: 75284298
Test: Manually with single and group notif
Change-Id: I57c3ac27e0c134ce04b681717b185eceba19a7a8

Former-commit-id: 56eb09158fdd1b8d9b5fe6718324abbc6555108a