Measuring the content views always at their specified size
Before we were limiting the height of the contentViews with
the height of the size that was given to us. When the
shade is collapsed, this actually made them measure
with size 0 in cases where we had a notch.
In rare cases, the measuring caching could lead to oddities
where content would be invisible.
We're now measuring them uniformly.
As a side effect, this measuring might also fix another bug,
where the notification is the full height of the view.

Test: add all sorts of notifications with notch, observe normal layout
Change-Id: I714f34f1564549952423ca51751cfe37ed45d991
Fixes: 73952748
Fixes: 78144261

Former-commit-id: be570ce9d7e0ddd33a92169870985938bac90d08