Prevent IME from targeting home when closing app is still animating
There were a few issues that caused the IME target to get set to home.
1. Home is considered above the other apps since it's getting launched.
However, the visual representation is actually below the closing app.
IME tries to target the highest app, but it's not necessarily the
highest visual app
2. computeIme was called before all apps were set up. This caused the
IME target to get recomputed before adding the closing apps
3. The condition for a WS.isClosing was based on previous versions so
it was incorrect

Test: IME closes on top of current target when going home with IME open
Change-Id: Ia9b0cc913eeb76a4af20f9ac7b667c5a3d1eafc7
Fixes: 78214125

Former-commit-id: 3a74855bbb160b2e9240c2268c823916af2dea72