Fix SntpClientTest failures with no active network
The last change is using getActiveNetwork to run the queries, which
results in an NPE if there is no network on the device. Using network 0
causes no socket tagging and allows queries to the local test SNTP
server to go through as previously.

Also migrating to JUnit4 and moving mServer and mClient initializers to
setUp().

Test: SntpClientTest now passes
Change-Id: Ieb0a5d247129bcad89c1add0e9c1c504f516e2a9

Former-commit-id: 6657d42602a25264ccbf70ce2eb36f3974518139