Remove raw coordinates VelocityTracker workaround
VelocityTracker uses raw coordinates starting in P, so it is no longer
necessary to adjust the MotionEvent to the screen coordinates in order
to get the correct velocities.

Bug: 72263561
Test: manual test - interaction with notification panel and quick
settings. Tried to fling it different ways (starting point, fling speed,
simple liftoff without flinging) and ensured that no unusual behaviour
is observed.

Change-Id: I19590ddeab1054a59b2671d6ef0ef7027509f863

Former-commit-id: 1db29f896d0e44e140c5861dde3ca2884b6f93f5