Status bar icons 15 dp
Shrink all status bar icons by 2dp. This means setting a scale factor on
anything that is a StatusBarIconView, and manually scaling down wifi and
mobile data groups.

Also fix the padding between the mobile data view and the battery

Also snuck in a fix for the battery saver icon clipping on QS

Change-Id: I5f5430d9e9c66f5c2794dcf4974cf18aa3bff061
Test: visual
Fixes: 74985733
Fixes: 79967193
Bug: 77822905

Former-commit-id: 39ea8100707a0581509c07d5f16fb813a2e311c2