Fix fullscreen notification to support RTL
In the FullScreen notification condition, HeadsUpStatusBarView
doesn't support RTL funcationality and show nothing on status bar.

There are three part need to support RTL:
* HeadsUpAppearanceController count translationX
* HeadsUpStatusBarView count the new padding start
* Consider more about the SystemWindow(Navigation Bar + cut out)

In normal case(LTR), to count the new padding start needs to know
both of the NavigationBar's width and cut out in the left hand
side. But, in RTL, to count the new Padding start needs to know both
of the NavigationBar's width and cut out in the right hand side.
To change from mLeftInset to mSysWinInset and from mLeftCutOutInset
to mCutOutInset.

Add HeadsUpAppearanceController.getRtlTranslation to count the RTL
tranlationX by the following formula:
  Y = screenSize - (SystemWindow's width + Scroller.getRight())
And, let HeadsUpStatusBarView.setPanelTranslation to handle the
right cut out width and then complete the real
HeadsUpStatusBarView.setTranslationX.

Bug: 79957423
Test: atest SystemUITests
Change-Id: I26b3cf8a73a2bcd2916847d72e66c16ca9063efd
Fix: 79957423

Former-commit-id: 3d00f35c758ce3b4f04f626e5711e6710cb1b8b0