Initialize mPendingAutoBrightnessAdjustment to NaN.
It currently defaults to 0 which is a valid adjustment, so if you have
a pre-existing adjustment it will get reset at every reboot and look
like a user adjustment.

Also, go ahead and set the pending adjustment to NaN if we've seen it
before so we don't try to apply it again in the future.

Bug: 80119319
Test: manual
Change-Id: I206f6d631ec2977a2c1a44f91e0c6c6a09ccd7ec

Former-commit-id: 9fdf66b0f19625be5d86d71089668f68eaf20890