Force-update the orientation of before sending to client
The changes in this topic focus on cases when we ensure visibility
and config when making some activity visible. Sometimes in WM we
might skip updating orientation because we're waiting for something
(checks marked with !forceUpdate), which results in configuration of
the previously visible activity being applied to the newly visible
one.

Bug: 76011287
Test: ActivityManagerAppConfigurationTests
Change-Id: I16b1de5ea9e38883692a4658f305518facaf98e9

Former-commit-id: 6cdcfe4e2b84fb8e2954bf69e3575aad811301ef