Pass correct preserve window value to performStop.
W/o passing correct preserve window value, performStop will mark the
window stopped as well, which consequently destroys all hardware
resources (display list etc.) associated with it. This causes flickering
when user drag resizes a task over a configuration change boundary that
needs to relaunch that activity, because the display list that draws the
content is cleared and not drawn.

Bug: 80320060
Test: Drag resizing doesn't cause flickering anymore.
Change-Id: I43fe5bf958ceb76817f6e8b9d1896d274eb40be1

Former-commit-id: c1681e6b1d62028ff6d0e8a7614587a22da08659