Notify WM that app is delayed closing due to the possibility of PIP.
WM determines the IME target after it sets up all the app transitions
and handles visiblity. In most cases, this is fine since everything is
set up properly at once. However, in cases where the app can enter PIP,
the transition to close the app is delayed in case the app decides to
enter PIP. In this case AM needs to notify WM of this information so it
can properly calculate the IME target when the app transitions aren't
fully set up.

The specific usecase is the following:
1. App with the ability to enter PIP is open with IME
2. Home is pressed.
3. Launcher is brought to the front, previous app is not yet hidden
since the close is delayed.
4. IME is recalculated and doesn't see any closing app. IME is
targeted to launcher.
5. Prevous app is ready to close.

By this point, the IME was already incorrectly targeting launcher, which
causes the flicker since IME will get set relative to launcher.

Test: Open Chrome with IME. Press home. No flicker.
Change-Id: Ie3553b0a01ad567d8b34c3453ad838b88684e79d
Fixes: 80352830

Former-commit-id: 4ad54910a1282c36ced893c4c2a010b32a30b920