Fixed an issue where the shadow wasn't rounded properly
Previously the implementation required the bottom roundness
to be treated specially when it came to the outline and
the clip path but this isn't necessary anymore and was
leading to issues with the shadows.
We now always apply the clipBottomAmount to the outline and
the clipping and can remove the special case.

Test: add notification, observe rounded clipping on bottom
Test: add groups, observe normal rounded clipping
Fixes: 80431867
Change-Id: Id65d11ce48c08adc138e1177226a03bc25ef6345

Former-commit-id: defd46f3a9b615694e674eb6d4ab12633f533f2f