Don't "notifyAppStopping" when going to sleep.
The WindowManager interprets notifyAppStopping as a signal that the
app surface will be destroyed, but in the visible stopped states
of going to sleep this may not necessarily be true. notifyAppStopping
was only recently introduced for the purposes of handling detach children
which isn't needed when turning the screen off. Simply not executing
it when turning the screen off should be safe, if a little confusing.

Test: Manual of bug + mess around with SV apps a bunch
Bug: 80491358
Bug: 72921025
Change-Id: I7dd5d1e144e12ecfba038e15f84f9ae7c02aadd2

Former-commit-id: 9e1bf7c7cca8df2173a5fe1c52ac16011051af89