After an ax service permanently crashes, it won't eat keys
The volume keys will work again, the ax button will reset back to the
correct state, and a bunch of other things.
This is cause we are not tracking the state of the service correctly in
this "Service won't run, but is on" state. Previously we just assumed it
was on.

Bug: 74593001
Test: Verified on device. By getting testback to crash. Tested, turning
service on and off in settings. Crashing it once, and having it restart
automiatically. Continuesouly crashing it until the system decides this
is bad. And after that state rebooting to make sure the service starts
up again.

Change-Id: I00e9c978d6807d8508968286a7545e48f9ae5800

Former-commit-id: fba2c8fb58412a314d6ce2c722d6614d21bdc99d