Lazily request spell checker session
Currently when an EditText is attached to a Window, one spell checker
session will be immediately requested even when that EditText is
empty.  This is not efficient, especially for notification with direct
reply.

With this CL, EditText does not request spell checker session while it
is empty.  Basically this doesn't cause any user-visible behavior
change.

Bug: 109701487
Fix: 37434131
Test: Run 'adb shell dumpsys textservices' to make that a new spell
      checker session is established not when a notification with
      direct reply is posted but when the user actually starts
      replying on that notification.
Change-Id: Iee04f4c0e2d248041d01c528344b191b9875d122

Former-commit-id: 6008106a64b48022997dc77dfd3c89dd5d3c5050