Fixed an issue where the contentHeight was wrong
The content height was using the regular topPadding which
factored in the darkAmount, but wasn't updating the contentHeight.
Since updating the contentHeight is pretty expensive, we are
now using the regular topPadding whenever the lock screen
isn't fully dark, which will make sure that everything is
properly positioned and not faded.

Change-Id: Ida9f29443efc1eb80924b0e755ea5338adfe23f7
Fixes: 110091926
Test: adb shell setprop debug.force_no_blanking true and observe proper layout after animations

Former-commit-id: f4b04ae0bb9a24faaaadb21c7e1c89e78a540fd0