Logging a wtf if alarm delivery gets stuck
If alarm delivery is stuck, detect and log a wtf. This will help us
catch relevant information sooner and also indicate how frequently this
happens. Otherwise this is something that can easily go unnoticed and is
usually hard to debug once the alarms are running as expected again.

Test: Builds, boots, existing alarm tests pass.
adb shell dumpsys alarm

Bug: 78560047
Change-Id: Ife881f2c9649b8dfd87ba05a425483375361da64

Former-commit-id: 0e4510df376eb15f1abcda78b9baf078f74f49b9