Do not disconnect existing VPN if seamless handover fails
If error happens during establishing of VPN, the cleanup logic tries to
disconnect the partially-created VPN network. At the moment this is not
working during VPN seamless handover since in this case no new network is
created and the existing VPN will be disconnected. Fix by not disconnecting
during the VPN handover case.

Test: cts-tradefed run commandAndExit cts-dev -m CtsHostsideNetworkTests -t com.android.cts.net.HostsideVpnTests
Bug: 110238496
Change-Id: I8cb2ab50596e61b0ac09e89eb1048816f00180d6

Former-commit-id: 307c525382d5f1454e37a0ae9d785fec71306264