Allow Cell Location for SYSTEM_UID and ROOT_UID
This change permits the system uid and root uid to
access cellular location information via the binder.
Previously this was restricted to the phone uid, but
running with uid=system is a privileged situation,
which makes me think this this wasn't intentional.

Also add a few lines of debug code to make issues
in LocationAccessPolicy easier to track down in the
future.

Bug: 110806860
Test: manual - ran with SL4A as SYSTEM_UID and
    verified access to getAllCellInfo.

Change-Id: Ie18be2cd72c49f1859d1434428f82f164bed8756

Former-commit-id: 2cdcd9837910e372a247907930deb50477186013