Reduce overview proxy connection backoff time to hide back faster
Switching to guest user would show a dialog to continue or start the
guest profile again and this would show the back button. At the same
time the connection is being established with the launcher again. First
time would sometimes fail causing a timeout of 5 secs before trying to
reconnect again. During the 5 sec timeout the back button would continue
to be visible until it reconnects and hides it. If this is reduced to 1
sec then reconnection would occur quicker and as a result hide the back
button.

Fixes: 109662462
Test: switch to guest user, immediately press continue and wait for back
button disappear

Change-Id: Id81b7f175834fb1eae8a0787674bbc130c91a5a2

Former-commit-id: 9608748b5ec30ce9a5a79e75bba9c4f724ebe9d3