Fix QS peek animation from ACTION_DOWN
When a notch is emulated, the EmptyShadeView gets measured as EXACTLY
60px, and this was causing
NotificationStackScrollLayout#getOpeningHeight to find the wrong value
for the height of the EmptyShadeView, thus the wrong amount to peek QS
open.

I'm not entirely sure why this isn't the case on devices without notch
emulation, but it seems that we should always let the children of
NotificationStackScrollLayout measure themselves as UNSPECIFIED so that
content heights are accurate while animating in.

Fixes: 110121269
Test: touch and hold status bar with no notifications; QQS should fully
expand when notch emulation is turned on.

Change-Id: Ie023c7ec813ea17d8378cc0f51e6aea5e3491722

Former-commit-id: eb7dbd5250d7664c7290e2b56a7495a604f5d8e9