Keep track of all bound services.
When a service is disconnected mServices will be updated
to no longer contain the service binding. In order to
prevent an accidental double-bind for services bound directly
in ManagedServices due to not detecting a service is still bound,
make sure that if the service is still bound that we don't rebind.

Bug: 77821801
Test: atest FrameworksUiServicesTests:ManagedServicesTest
Test: manual - check service bound after force-stop/app update
Test: manual - check service not re-bound after disabling a component

Change-Id: Id600357ad0b84f157cf9666680bf53b7fe125d96
(cherry picked from commit d814134e9ee652de5b54c7c20c8eb3c1a8f590df)

Former-commit-id: 190c06736a4affd83fdcc77590faf94ac5a20c2c