fix overlay on status bar because of two clock
Although the status bar block user's tap, the user still could
drag the status bar to expland the notification panel.

The root cause is that neither the clock nor the notification icon
inner area doesn't hide when the user drag down the status bar in
landscape condition but portrait does.

The solution is to trigger mStatusBar.recomputeDisableFlags when
the heads up is showing in landscape mode and the expand status is
different. CollapsedStatusBarFragment.adjustDisableFlags should
disable clock showing when the heads up showing.

Change-Id: If7411f5d8eef9cf234b83a463006325ba0156311
Fix: 110808257
Test: atest SystemUITests

Former-commit-id: ca4eee83d9d7e920cc01e5762f6d9056bbe3f546