Menu ime button takes nav button width to show recents button for tablets
Menu ime button was pushing the recents button out of screen because of
the order of the nav layout (where right side was before the recents
button) and menu ime button took full width (match parent). Fixed with
a fixed width for menu ime button to show the recents button on nav bar
for tablets (screen density smaller than ~250).

Change-Id: I61a0206e40b7a69d52db4a97a1b76767ba752f85
Fixes: 77913318
Test: 'adb shell wm density 240' on phone

Former-commit-id: eb9bf01fd879fa29049ceb5b8b6a74a9f916c928