Fix the lost RAM calculation
The formula used when reporting memory usage when device goes
low on memory should exclude SwapPss, which was incorrectly
being included as part of totalPss.

BUG: 109762356
Test: Manually tested on a gobo device by making it
go low on memory. Verified that the Lost RAM in adb logcat
was negative before, and is positive with the change,
and close to Lost RAM from "dumpsys meminfo".

Change-Id: Ic54181b55fa03d3faebda4c0516c95c18ee39e41

Former-commit-id: 786d5453bfd73f9ee740f6ead12b0514f7e6023c