Add API to retrieve Audio Product Strategy from a given Audio Attributes
In order to ensure only one matching algorithm is implemented in the
audio policy engine, we cannot use the list of audio product strategy
to let the client of this API inferring the product strategy to use.
It shall request the audio policy manager which product strategy shall be
followed by a given Audio Attributes.

Bug: 124767636
Test: dumpsys media.audio_policy
Change-Id: I372a47df4a35efd68a3ea327581af54823997dba
Signed-off-by: François Gaffie <francois.gaffie@renault.com>

Former-commit-id: f4f3bfe74779601c5e240892c847b785342d4901