Send interaction hint when starting window animation
If we are starting a window animation, there is a very strong
correlation that usually just before the user took some kind of
action.

However, in some cases the power hint may be lost of there are not
continiously produced frames between the user interaction and the
window animation, like when opening the IME but it takes a while
until the IME comes up, or when the status bar automatically hides
after being temporarily shown in a fullscreen app.

Test: SurfaceAnimationRunnerTest
Test: Go to an immersive_sticky app, swipe down from top, wait
until status bar hides, make sure interaction hint gets sent.
Bug: 111253599

Change-Id: Id517d48a63d4c528265a808c49043d936d804625

Former-commit-id: b6ba4334b46c378a7a8480de023e04f8079b2dd3