passpoint-r1: fix the wrong filed for PPS-MO.
Currently Passpoint-r1 follows the PPS-MO format to install passpoint
profile.

Howevever, one field(SubscriptionParameter) is not correct.
Normally passpoint-r1 doesn't use this field because this is used for
passpoint-r2 feature.
So, currently there is no issue with wrong field for using passpoint-r1.

But, it needs to be "SubscriptionParameters" according to the hostpot2.0
release 2.0 techinical specification.

Bug: 111449001
Test:  ./frameworks/base/wifi/tests/runtests.sh
Change-Id: Idb6bb4d28913bd3d0905b330e25fbb20b99bbf2e
Signed-off-by: Ecco Park <eccopark@google.com>

Former-commit-id: 84d601fbda7bee7ffecfc5d8589b710495846681