Parcel mCharacters in KeyEvent
Currently, the mCharacters field gets lost via IPC because it is not
written to Parcel during serialization. Add this field to the parcel.

The ACTION_MULTIPLE with KEYCODE_UNKNOWN signals usage of mCharacters.
But the mCharacters field isn't really used by the system.
So, deprecate this field.

Also deprecate ACTION_MULTIPLE, which doesn't really get used.

Bug: 111386868
Test: atest cts.KeyEventTest
Change-Id: Ie01b796a10f8b2dcc2bcfa15f71b17a7e917faa7

Former-commit-id: e481d33cc0f02ff63d337304b9d4e3bddf42179a