When quickstep resets, set the nav bar alpha back to 1f
When rotating the screen, the animation to animate the nav bar back from
a quickscrub track would not finish animating to 1f because the reset
sets the current nav to null and leaves it transparent. Force set the
nav bar to 1f after will make sure it is opaque.

Test: in landscape use portrait forced app and quick scrub to another
app without forced orientation, and go back and forth
Fixes: 111303134

Change-Id: I632657fec07d721e4ee3f41624de67e3e1b4de29

Former-commit-id: 819daaa5e7bbdfa09dd1c918106c66546a54dd6c