Send widget registration broadcasts before BOOT_COMPLETE
Widget presence in the home app is important for usability, but it
depends on a broadcast-based handshake at boot time.  This handshake
occurring after the BOOT_COMPLETED broadcast was initiated means
that in practice widgets may not become available for literal
minutes following unlock, as it can take this long for the
boot-complete broadcast to clear and let the systen proceed with
dispatch of the widget handshakes.

We address this by hoisting the widget setup broadcast to occur
just *before* the boot-completed broadcast, rather than as part
of general listener reaction to the global "this user has been
unlocked" notification.

This is a manual-merge patch to permit upstream submission of a
minor fix with the substantive CL.

Bug: 76154638
Test: manual (note broadcast ordering following boot)
Change-Id: Ibd97268f995ec673f21d8f5df257041cf61a058d

Former-commit-id: 9046e3d400744100c784cceabb025e549d3e2096