RedactingCursor that redacts specific columns.
For new storage changes in Q, we need to redact the visibility of
the "_data" column on any returned cursors.  We could cook this up
using SQLiteQueryBuilder projection re-mappings, but that's a pretty
involved change due to the number of tables and teasing apart
public-vs-hidden API columns.

Instead, this RedactingCursor class offers to "redact" requested
columns by replacing with specific values, such as "/dev/null".

Test: atest frameworks/base/core/tests/coretests/src/android/database/RedactingCursorTest.java
Bug: 111960973
Change-Id: Iad36986fc0b7a0cae10ff216755e284320991e18

Former-commit-id: a5d0bf17013198afbfe2671967ff905153f66e07