Unclone autofill service setting in managed profile
The autofill service setting was cloned from the personal
profile to the managed profile in ag/2126240 to fix a bug where the
Settings UI did not control the managed profile. That fix caused
unintended interactions between both profiles. The setting has
now been 'uncloned' and a separate UI control introduced for each
profile in ag/4666600.

BUG: 3803359
Test: Tested manually by setting up a work profile and verifying that
the setting can be changed independently.

Change-Id: I7b5cac69cb8b4b0782538cb86f818904d6ad238d

Former-commit-id: 9822ffc49e2d83f3c6cb91d835e8148b4fce649f