Use the current oriented button to the nav button drawable's callback
In ButtonDispatcher, when setting a new drawable, it would apply it to
each of the views (orientations) for the same drawable, unfortunately
the last view that was set sets the callback of the drawable to itself.
This causes any invalidations to the drawable to invalid the wrong view
because the callback was attached to the last view set. Fixed by
applying the current view to the drawable when setting the drawable or
current view.

Bug: 111907556
Test: manual, log drawable's callback when invalidated
Change-Id: Ifa10b86555047116637eedae50df78c1173d6c3a

Former-commit-id: b007c774641a991bbb7c35de9248e95582826185