Add "owner_package_name" column.
To support new storage features in Q, we need to start tracking the
"owner" of each inserted file.  This change introduces the schema
change to insert the column, including logic to derive values for
existing well-known package-specific paths.  Populate the column for
any newly inserted media based on the calling package.

Bug: 111893193
Test: adb shell am instrument -w com.android.providers.media.tests
Change-Id: Ibf5656423c22e7e540ab4a3a3efb99b8af944e16

Former-commit-id: 0ce83aeb2b04a279380f062a07c46c19eccb3785