Pinner: Don't pin apps for work profile.
Since the Home and Camera intents aren't set to relevant apps for the
work profile, we don't want to pin those. However, the pinner will
trigger on boot for the work profile, after the primary profile has
pinned, and clear the currently pinned apps to try to pin the work
profile apps. This change skips rerunning the pinner for the work
profile.

Bug: 112357464
Test: Added logging and tested device with a work profile. After reboot,
saw that the apps for the primary profile are pinned and the ones for
the work profile are not. Previously, apps for the work profile were
attempted to be pinned.
Change-Id: I0a7fdeab4287dcb6765c0fb47647ddc36db405b6

Former-commit-id: 0ff9946abe9569808fb6b559aefb0c77f514c5ec