ResourceManager: Disable APK Assets cache.
The WeakReference based cache of Assets in use elsewhere in the
process remains unchanged.

Note that this change is larger than it needs to be because it is
an error to create an LruCache of size zero.

There seems to be no measurable difference as measured by the standard
app startup benchmarks.

Test: atest -v google/perf/app-startup/benchmark-app-hermetic/cold-dropcache-test
Bug: 111385832
(cherry picked from commit c5448f2138692b472ad6aa7e47725baff17ca342 [formerly bc95621811b0c15904a4783aae34e94aa109f764])

Change-Id: Ib7a52bd03f6f13cdb76b82abc307d17227c4c1dc

Former-commit-id: 77e9589274f85e30d68affd3f3defc109ae1041f